+++ lib/local_db/product/tests/test_productparameter_data_source.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:openflutterecommerce/local_db/product/data_sources/product_parameter_data_source.dart';
-import 'package:openflutterecommerce/local_db/product/entities/product_parameter_entity.dart';
+import 'package:openflutterecommerce/local_db/product/entities/productparameter_entity.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
